More expressive diagnostics in error cases (#21)
JuliaParser had the ability to give detailed error reports
when parsing failed. Start porting a couple of these diagnostics
over to CSTParser. The new `FancyDiagnostics` package can be used
to play with this interactively:

```
julia> abc(d
REPL[4]:1:6 ERROR: Unexpected end of input
abc(d
     ^

julia> a && && b
REPL[5]:1:6 ERROR: Unexpected operator
a && && b
     ^~~
```

This doesn't port the entire suite of JuliaParser diagnostics, just
a couple to be useful. However, this should get the basic infrastructure
in place upon which we can iterate.