Drop errors.jl file for now
I missed the existence of this file and added error codes to hints.jl.
It doesn't look like these are really used, so drop that file for now.
We might want to move the error codes out of hints.jl at some point,
at which point we can re-create this file.