Whitespace refactor (#32)
* Start whitespace refactor

No functional changes in this commit, but should make subsequent changes
easier.

* Minimize use of span field

In preparation for a richer representation of whitespace.

* Introduce `fullspan`

The newdesign is as follows: `fullspan` is what `span` used to be
and includes all leading and trailing whitespace. `span` is now the
location of the actual (non-trivial) part of the expression. This
commit just plumbs through the changes. No real effort at correctness yet.

* Use AbstractTrees for display purposes

* Keep punctuation in string interpolation syntax

For refactorings all characters need to be accounted for. These were
not, so add some more syntax nodes to account for this punctuation.

* Fix a number of incorrect spans

* One more span bug

* Remove linting and variable definitions (#1)

* fix `mutable` span (#2)