Better support for :incomplete errors (#29)
* Drop errors.jl file for now

I missed the existence of this file and added error codes to hints.jl.
It doesn't look like these are really used, so drop that file for now.
We might want to move the error codes out of hints.jl at some point,
at which point we can re-create this file.

* Better support for :incomplete errors

Base parsing distinguishes incorrect syntax and incomplete syntax.
For the latter, pressing enter at the repl will give a newline rather
than a parse error. Add the appropriate error kinds and try to emit
them as necessary. FancyDiagnostics has support to hook this into Base.