Support bracketed function call in function definition
The following syntax:

```julia
function (f(a, b))
end
```

was interpreted as an anonymous function with 1 argument of name `f(a,b)`.
This commits allows this syntax as a function `f` with two arguments.

This happened to me in the following case where the references to `a`
where resolved as global `a` (instead of as an argument):

```julia
function (a + b)
    a
end
```